Updated Scaffold A
Made it check all block faces, horizontally and vertically. This should stop expand or downward scaffolds.

Fixed a possible false in vehicles, inside of block and against doors.
Fixed a possibly rare false flagging on blockplace instead of waiting for the next flying packet.
Added comments to make it easier to understand as a resource for newcomers. :D